Pretty extensive changes:
Changed Graphics display style to have ship in center and draw everything relative to it
   New Star internal class to store the background, to give a sense of movement.
   New Message internal Class to store messages written to the screen

New Weapon logic:
   New Weapon and WeaonManager interfaces; Missile implements Weapon and Missile Manager implements Weapon Manager.
   Most (if not all) methods involving "Missile" now renamed with "Weapon".
   Changed Missiles to now accelerate (with Particles, subject to lag) when fired, more like real Missiles.
   Added Bullet and BulletManager class.  Bullets differ from Missiles in the fact that they do not accelerate and can only hurt one Asteroid.
   Bonuses are now mostly added and removed through WeaponManagers.
   New enum in Ship: WeaponType.  Starts with Missiles, but can be toggled with 'Q'.
   Shooting now works like moving, with a boolean in Ship of isShooting, instead of individual calls to shoot().



Added a FPS counter.  Right now it just does a System.out.println(), but that can be easily modified.
Negligable changes to ParticleManager:  cleaned up a javadoc and removed an unnecessary statement of i += 0.
Moved Ship particle code to generateParticles().
Moved setMaxShots() and getMaxShots() to WeaponManagers.

Still need to add documentation and fill out the boni in BulletManager.