Started to make a subclass of Alien called PredatorAlien... I still need to look into your coordinate system more before I can write the methods that actually move the Alien in predatory behavior.
PredatorAlien's generic behavior: Finds a prey (ship) in its designated radius of scope, accelerates toward it until it reaches a certain proximity then proceeds to attack it. Will loose the prey if the prey moves out of the Predator's radius of scope. 

Changes to Alien: divided and conquered the act method and grouped the more Alien unspecific code to a protected method named generalActBehavior(). This was a work around for the PredatorAlien subclass so that I didn't have to un-encapsulate much.

Anyway, I hope I did more good than harm!