NEW WEAPON!!
Added the little doctor, based on the weapon of the same name from Enders Game
The graphics of this weapon could be improved, and the constants that determine its working can be changed, but the framework is there.
The weapon needs to be checked for syncing, but it shouldn't be too far off (mostly just coppied Missile)

Aliens changed so that you can't crash the game by shooting them with the Lil' Dr (they no longer check collisions if they're dying)

Guided Missiles now look for their target where they will be, instead of where they are, to avoid some abrupt direction changes.
To that end, added projectX(time) and projectY(time) to GameObject.