A lot of work to the Guided Missile
Completely redid the way it looks for a target; could use some tweaking on the comparison method.
Added the turn method to GameObject (see javadoc)
Using said turn method, redid how Guided Missiles track their prey.  It looks a lot better as long as the Missiles don't fly in circles.
Added the Guided Missile as a weapon in the ship's arsenal.